+++ packages/devtools_app/lib/src/shared/analytics/constants.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:devtools_shared/devtools_extensions.dart';
+
 import '../screen.dart';
 
 part 'constants/_cpu_profiler_constants.dart';