+++ packages/devtools_extensions/example/foo/packages/foo_devtools_extension/lib/src/foo_devtools_extension.dart
@@ -2,9 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:devtools_extensions/api.dart';
+import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_extensions/devtools_extensions.dart';
 import 'package:flutter/material.dart';
+
+import 'devtools_event_example.dart';
+import 'devtools_extension_api_example.dart';
+import 'expression_evaluation_example.dart';
+import 'service_extension_example.dart';
 
 class FooDevToolsExtension extends StatelessWidget {
   const FooDevToolsExtension({super.key});
@@ -25,38 +30,6 @@ class FooExtensionHomePage extends StatefulWidget {
 }
 
 class _FooExtensionHomePageState extends State<FooExtensionHomePage> {
-  int counter = 0;
-
-  String? message;
-
-  @override
-  void initState() {
-    super.initState();
-    // Example of the devtools extension registering a custom handler.
-    extensionManager.registerEventHandler(
-      DevToolsExtensionEventType.themeUpdate,
-      (event) {
-        final themeUpdateValue =
-            event.data?[ExtensionEventParameters.theme] as String?;
-        setState(() {
-          message = themeUpdateValue;
-        });
-      },
-    );
-  }
-
-  void _incrementCounter() {
-    setState(() {
-      counter++;
-    });
-    extensionManager.postMessageToDevTools(
-      DevToolsExtensionEvent(
-        DevToolsExtensionEventType.unknown,
-        data: {'increment_count': counter},
-      ),
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -64,53 +37,19 @@ class _FooExtensionHomePageState extends State<FooExtensionHomePage> {
         backgroundColor: Theme.of(context).colorScheme.inversePrimary,
         title: const Text('Foo DevTools Extension'),
       ),
-      body: Center(
+      body: const Padding(
+        padding: EdgeInsets.symmetric(vertical: denseSpacing),
         child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
           children: <Widget>[
-            Text('You have pushed the button $counter times'),
-            const SizedBox(height: 16.0),
-            ElevatedButton(
-              onPressed: _incrementCounter,
-              child: const Text('Increment and post count to DevTools'),
-            ),
-            const SizedBox(height: 48.0),
-            Text('Received theme update from DevTools: $message'),
-            const SizedBox(height: 48.0),
-            ElevatedButton(
-              onPressed: () => extensionManager
-                  .showNotification('Yay, DevTools Extensions!'),
-              child: const Text('Show DevTools notification'),
-            ),
-            const SizedBox(height: 16.0),
-            ElevatedButton(
-              onPressed: () => extensionManager.showBannerMessage(
-                key: 'example_message_single_dismiss',
-                type: 'warning',
-                message: 'Warning: with great power, comes great '
-                    'responsibility. I\'m not going to tell you twice.\n'
-                    '(This message can only be shown once)',
-                extensionName: 'foo',
-              ),
-              child: const Text(
-                'Show DevTools warning (ignore if already dismissed)',
-              ),
-            ),
-            const SizedBox(height: 16.0),
-            ElevatedButton(
-              onPressed: () => extensionManager.showBannerMessage(
-                key: 'example_message_multi_dismiss',
-                type: 'warning',
-                message: 'Warning: with great power, comes great '
-                    'responsibility. I\'ll keep reminding you if you '
-                    'forget.\n(This message can be shown multiple times)',
-                extensionName: 'foo',
-                ignoreIfAlreadyDismissed: false,
-              ),
-              child: const Text(
-                'Show DevTools warning (can show again after dismiss)',
-              ),
-            ),
+            ServiceExtensionExample(),
+            SizedBox(height: 32.0),
+            EvalExample(),
+            SizedBox(height: 32.0),
+            ListeningForDevToolsEventExample(),
+            SizedBox(height: 32.0),
+            CallingDevToolsExtensionsAPIsExample(),
           ],
         ),
       ),