+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,6 +1,18 @@
 ## 0.0.10
+* Bump minimum Dart SDK version to `3.3.0-91.0.dev` and minimum Flutter SDK version to `3.17.0-0.0.pre`.
+* Add a test target to the `app_that_uses_foo` example that can also be debugged
+with the DevTools extension provided by `package:foo`.
+* Add an example of performing expression evaluations from a DevTools extension.
+* Add an example of registering a service extension and calling it from a DevTools extension.
+* Document the DevTools extension examples.
+* Add documentation to [ExtensionManager] public APIs.
+* Fix some bugs with the `build_and_copy` command for Windows.
+* Add an example `launch.json` file in the `example/foo` directory.
+* Clean up the package readme to make instructions Windows-compatible.
 * Update the README with instructions for joining the Flutter Discord server.
-* Bump the `devtools_shared` dependency to ^5.0.0
+* Bump `package:devtools_shared` dependency to ^6.0.1
+* Bump `package:devtools_app_shared` dependency to ^0.0.7
+* Bump `package:vm_service` dependency to ^13.0.0.
 
 ## 0.0.9
 * Add a link to the new #devtools-extension-authors Discord channel in the README.md.